requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
opencv-python>=4.5.0
tqdm>=4.60.0
matplotlib>=3.3.0
pandas>=1.3.0
pyyaml>=5.4.0
pillow>=8.0.0
